@@ -17,7 +17,8 @@ module.exports = function (context, options) {
           </style>
           <div class="footer footer--${options.footerStyle} docusaurus-dynamic-copyright">
             <div class="container text--center">
-              ${options.copyright}
+              ${context.i18n.currentLocale === 'en-us' ? options.copyright : options.copyright2}
+              ${context.i18n.currentLocale === 'en-us' ? '<p>Official Address: 4711 Yonge St, North York, ON M2N 7E4, Canada</p>' : ''}
             </div>
           </div>`,
         ],
