@@ -1,5 +1,4 @@
 import React from 'react';
-import {useEffect} from 'react';
 import clsx from 'clsx';
 import Layout from '@theme/Layout';
 import Link from '@docusaurus/Link';
@@ -9,7 +8,6 @@ import styles from './index.module.css';
 import HomepageFeatures from '../components/HomepageFeatures';
 import hldHdImage from "../../static/img/SRS-SingleNode-4.0-hd.png";
 import hldSdImage from "../../static/img/SRS-SingleNode-4.0-sd.png";
-import useIsBrowser from "@docusaurus/core/lib/client/exports/useIsBrowser";
 
 function HomepageHeader() {
   const context = useDocusaurusContext();
@@ -24,9 +22,6 @@ function HomepageHeader() {
         <div className={styles.buttons}>
           <Link className="button button--secondary button--lg" to="/docs/v6/doc/getting-started">
             {translate({id: 'homepage.getStarted'})}
-          </Link>&nbsp;
-          <Link className="button button--secondary button--lg" to="/cloud">
-            {translate({id: 'homepage.cloudService'})}
           </Link>
         </div>
       </div>
@@ -42,37 +37,6 @@ function HomepageHeader() {
 }
 
 export default function Home() {
-  const isBrowser = useIsBrowser();
-
-  useEffect(() => {
-    const loadGlimeScript = () => {
-      const script = document.createElement('script');
-      script.src = 'https://cdn.glimelab.ai/widget/1.0.0/widget.js';
-
-      script.onload = () => {
-        // Call the window.glime.init function here
-        window.glime.init(`odGnkiPMumIMJp`, { styles: { baseFontSize: 16, theme: "dark" | "light" | "auto", hideCollapsedButton: false } });
-      };
-
-      document.head.appendChild(script);
-    };
-
-    if (isBrowser) {
-      loadGlimeScript();
-    }
-
-    return () => {
-      // Clean up any resources when the component unmounts
-      if (isBrowser) {
-        // Remove glime script if added to the DOM
-        const script = document.querySelector(`script[src="https://cdn.glimelab.ai/widget/1.0.0/widget.js"]`);
-        if (script) {
-          script.remove();
-        }
-      }
-    };
-  }, [isBrowser]);
-
   return (
     <Layout
       title={`SRS (Simple Realtime Server)`}
@@ -82,7 +46,7 @@ export default function Home() {
       <main>
         <HomepageFeatures/>
       </main>
-      <img src='https://ossrs.net/gif/v1/sls.gif?site=ossrs.io&path=/lts/index'/>
+      <img width={0} src='https://ossrs.net/gif/v1/sls.gif?site=ossrs.io&path=/lts/index'/>
     </Layout>
   );
 }
